CSPACE-4942: more styling fixes: columns, repeatable widget, calendar widget
CSPACE-4942: fixed column width bug where .info-pair* was set to 100%;
added class for tables containing the repeatable widget so that the
primary and delete inputs are better positioned (so far only in
CatalogingTemplate.html. needs to be extended to other template files.
could be added via JS…); modified the padding for drop down select
elements; modified the margin-top property for the calendar widget