NOJIRA: Changes to sign on page text
PAHMA-CollectionSpace 2.4

Welcome to PAHMA's new collection management system, CollectionSpace.
PAHMA-CSpace has been configured to support the needs of anthropology
museums in general, as well as the specific needs of the Phoebe A.
Hearst Museum of Anthropology.

To sign in, enter your email and passphrase in the Sign In box on the
right
.
If you do not yet have access to PAHMA-CollectionSpace, have your
supervisor send a request on your behalf to pahma-cspace@berkeley.edu
with your specifics as well as the levels of access that you'll be
needing.